requirements.txt
aniso8601==7.0.0
astroid==2.7.2
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-GraphQL==2.0.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
idna==3.2
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
platformdirs==2.2.0
pluggy==0.13.1
promise==2.3
py==1.10.0
pylint==2.10.2
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.19.0
requests==2.26.0
Rx==1.6.1
six==1.16.0
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
